+++ quaddtype/tests/utils.py
